additional round of experiments. need to figure out why fails
